Added Group one function control, 128 speed step control.
Also reworked the controller and the encoder so most of the time we only send
idle packages, which solves the problem of slow payload upgrades etc. When
a new payload is built, we proceed to send it.

This fixes a lot of issues and gives us a lot of time to process payloads
updates. The controller will send they payload every 4-5 seconds, but
in the meantime the locos will receive nice idle packages.

Added nice __repr__ methods to controller and locos.

Cleaned up and fixed many small issues. Tested extensively with 3 locos at the same
time